requirements.txt
Django==5.0.1
django-ckeditor==6.7.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
gunicorn==21.2.0
pillow==10.2.0
psycopg2==2.9.9
python-decouple==3.8
six==1.16.0
whitenoise==6.6.0
